Fix decross desync on next turn start.  More logging around states starting and ending. CustomBarProgress is now reset in CardSelectBattleState::onStart instead of the CombatBattleState.  Version bump.
The desync fixed was one where being decrossed during the turn and then not choosing a form on the following card select would cause the remote player to think you changed forms to the form at -1, which would animate and desync.

SetCustomBarProgress call moved in order to prevent progress from resetting during transitions between combat and transformation state during combat, ,such as through decross. Transformation state should not be added as a subcombat state, so this makes more sense.